Update SaveData.java
Use Java 8's LocalDate and LocalTime: Instead of using java.util.Date for date and time representation, it's advisable to use Java 8's LocalDate for dates and LocalTime for time. These classes offer a more comprehensive API and are more intuitive to use. This suggestion aligns with modern Java best practices, as mentioned in the Java Coding Best Practices article.
https://rhamedy.medium.com/a-short-summary-of-java-coding-best-practices-31283d0167d3