Update SavesService.java
Use Interface for Collection Types: Instead of defining the allSaveData member variable as an ArrayList, it's a better practice to define it as a List. This provides flexibility in changing the concrete implementation later without affecting other parts of the code. This suggestion is in line with the principles of coding to interfaces, a concept reinforced in the Java Coding Best Practices article.
https://rhamedy.medium.com/a-short-summary-of-java-coding-best-practices-31283d0167d3